[Xamarin.Android.Build.Tasks] Uninformative ""aapt.exe" exited with code 1." and "The file "obj\Debug\android\bin\packaged_resources" does not exist" errors if drawable resource filename contains invalid character
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=36485

Under some conditions apps produces an error which does NOT
contain "error <something>:". As a result our regex was not picking
up some errors and was just printing the default error of

	"aapt.exe" exited with code 1

This commit updates the regex to allow for errors which do
not include "error <something>:". And adds a unit test for the
specific test case.