[Xamarin.Android.Build.Tasks] Improve Aapt error message handling (#703)
Fixes: https://bugzilla.xamarin.com/show_bug.cgi?id=36485

Under some conditions apps produces an error which does NOT
contain "error <something>:". As a result our regex was not picking
up some errors and was just printing the default error of

	"aapt.exe" exited with code 1

This commit updates the regex to allow for errors which do
not include "error <something>:". And adds a unit test for the
specific test case.