[Xamarin.Android.Build.Tasks] Rework GetAdditionalResourcesFromAssemblies to not download in $(DesignTimeBuild)
Context https://bugzilla.xamarin.com/show_bug.cgi?id=57281

One of the problems we have with downloading content is the DesignTimeBuilds.
We CANNOT allow a DesignTimeBuild to download stuff from the internet.
It will lock up the IDE for however long it takes to get the data
and the user will have no indication as to what is going on.

That said.. in order to get intellisense working we need to make
sure that a default Xamarin.Forms project will at least build in
DesignTime mode.

So we have a problem. We cannot download stuff, but we need it to build.
Fortunately commit 4f10438c allowed the GetAdditionalResourcesFromAssemblies
Task to look in the android-sdk/extras folder to get its required
resources. So we can alter the Task to allow that part to happen, this
will extract the data to the correct location. But we will still make
sure that we do not download any additional data from the internet.

So providing the m2repository bits are installed in the android-sdk/extras
these changes will allow those to be extracted in a DesignTimeBuild.
The idea is those extras should be installed as part of the main installer,
which will allow the default project to build. Obviously if those extras
are not installed the build will still fail. But as already stated
that cannot be avoided.

We also added $(DesignTimeBuild) to the property cache files so we can
ensure that when we do switch to a main build we run GetAdditionalResourcesFromAssemblies
again.