[api-merge] change default interface methods to non-default when merging
... with old API that had non-default methods.

We cannot have default interface methods in C#. They have to be removed
until we get proper DIM support. And changing methods that existed in the
earlier API to default is a breaking changes in the resulting C# API
because those methods are simply removed.

To avoid that issue, change them to non-default if there was an existing
method.