[Mono.Android] remove AnnotatedElement.isAnnotationPresent() hack.
Now that we have appropriate fix in api-merge to not leave default interface
methods as is, remove the relevant workarounds in metadata.

This however introduced a build breakage at Parameter class because this
"new in API Level 26" class did not implement this exact method.
Until we get appropriate support for DIMs, add workaround manual
implementation based on what ojluni in Android framework exactly does.