[Xamarin.Android.Build.Tests] Fix DesignTimeBuild Test
The test was using the global cache to store the downloaded files.
As a result if we get multiple tests/commits building at the
same time we end up deleting files half way through a test run.

So lets make use of a local cache for the test by using the
`XAMARIN_CACHEPATH` environment variable.