[Xamarin.Android.Build.Tasks] android:debuggable="true" & RELEASE (#583)
Fixes: https://bugzilla.xamarin.com/show_bug.cgi?id=56075

`//application/@android:debuggable` should not be `true` set when
packaging apps in the Release `$(Configuration)`.

Looks like there was a change to either the IDE or the template
for VS2017. Rather than leaving DebugType blank or not setting it
in the new project it gets a value of 'None'. As a result our code
which decides if this is a debug build, which only checks for ''
fails.

This commit adds a check for "None" and adds the appropriate unit tests.