[Xamarin.Android.Build.Tasks] packageName replacement for Libraries (#647)
Context: https://bugzilla.xamarin.com/show_bug.cgi?id=56690

There is an issue in the current way we merge manifest document
`android:name` values. We do NOT process the `.foo` style name
values into `{packagename}.foo` . The component team currently
do this manually. It really should be done in our toolchain,
since users will probably hit this.

This commit adds a new step when merging in the library manifest
elements to fix up those values. It also adds a unit test to
check that we not only fix up those values for library projects
but also fix up the [`${applicationId}` place holder][1].

[Android documentation on this topic.][2]

All of the attributes which allow the replacement of a `.` with the
manifest package name have been included in this PR.

[1]: https://bugzilla.xamarin.com/show_bug.cgi?id=56690
[2]: https://developer.android.com/guide/topics/manifest/manifest-intro.html