[Xamarin.Android.Build.Tasks] Don't package reference assemblies (#706)
Fixes: https://bugzilla.xamarin.com/show_bug.cgi?id=57342

The investigation in the bug suggests that we are picking up a
reference assembly rather than a the actual implementation.
This is to do with the way netstandard nuget packages work, they
include both `ref` and `lib` folders.

In this case `ref` was being included in the package rather than
`lib`. This commit alters the `_ResolveAssemblies` to use

	@(ReferenceCopyLocalPaths)

This ItemGroup is populated with the correct items.