[xa-prep-tasks] fix bad Array.Copy() call (#722)
On Windows, `Array.Copy()` will throw an `ArgumentException` for various Git
MSBuild tasks. Turns out the call is using the wrong variable: the
string instead of the array. I am not sure why this works on macOS, I
suspect Mono isn't validating input the same way as Windows and it
_happens_ to work.