[Xamarin.Android.Build.Tasks] [VSM][Android] Build error observed with MultiDex + Custom configuration have spaces.
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=58134

We did a ton of stuff to get multidex working correctly on windows
in commit 6829b7d1. However it seems that if there are spaces in the
path on MacOS, we get similar problems.

The solution is to add single quotes around the list of .jar files
when calling the MainDexListBuilder. That seems pretty logical.
But... it breaks it on windows (sigh). Because on windows if we add
the single quotes, it stops working.

So we have to conditionally add the single quotes around the list
of jar files. Not this only applies when we call the MainDexListBuilder.
The single quotes seem to work ok when calling proguard..

Weird.

This commit also adds a unit test for this. However we needed to be
able to change the names of the Configurations "Debug", "Release".
So contructor arguments have been added to allow them to be overridden.
These will default to "Debug" and "Release" so that existing tests will
work as expected.