[Xamarin.Android.Build.Tasks] _CleanGeneratedDebuggingFiles target removes ALL .pdbs
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=58646

Rather than just blindly delete ALL the debug files in the
`$(OutputPath)` we should be a bit more selective.

So instead we add the files to the

	`$(IntermediateOutputPath)$(CleanFile)`

so that they can be cleaned up automatically using the in
build xbuild/msbuild Clean systems.