[Xamarin.Android.Build.Tasks] IDE0006 warning on File -> New CrossPlatform App (#730)
Context: https://bugzilla.xamarin.com/show_bug.cgi?id=58448

Based on feedback from the VS team we now know that library
projects are NEVER built as part of a DesignTimeBuild :(.

Furthermore, if ANY error occurs during the DesignTimeBuild
VS throws a wobbly and does not initialise the project
correctly. The suggestion is to use `ContinueOnError` to
ignore the errors and allow the build to complete.