[Xamarin.Android.Build.Tasks] Restore @(AndroidEnvironment) (#742)
We have a failure within
`EnvironmentTests.EnvironmentVariablesFromLibraryProjectsAreMerged`
(commit ea6b9b45): the `@(AndroidEnvironment)` from the library
project is not merged into the App `.apk`:

	Expected: null
	But was:  "Well, hello there!"

(The arguments were in the wrong order. Oops.)

The cause for this is due to `<ResolveLibraryProjectImports/>`
attempts to avoid duplicate work; [from the build log][0]

[0]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android/526/consoleText

	Task "ResolveLibraryProjectImports"
	  Assemblies:
	    .../xamarin-android/bin/TestDebug//LibraryResources.dll
	  Skipped resource lookup for .../xamarin-android/bin/TestDebug/LibraryResources.dll: extracted files are up to date
		ResolvedEnvironmentFiles:

*Because* `LibraryResources.dll` was skipped, the
`@(AndroidEnvironment)` entries from `LibraryResources.dll` were also
skipped! Doh!

Update the `<ResolveLibraryProjectImports/>` task so that when
resource extraction for an assembly is skipped, we remember to add the
already-extracted `@(AndroidEnvironment)` values.