[Xamarin.Android.Build.Tasks] [Smoke][VS 2017] Exception on deploying CheeseSquare sample
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=58721

The initial problem seems to stem from the fact that for the
support library components now directly include the resources.
As a result the `_GenerateJavaResourcesForComponent` was not
running. This was because it was only running if
`_AdditonalAndroidResourceCachePaths` was not empty.
In this scenario it was always empty.

As a result none of the files like R.java files were generated
for the support libraries. The means the final .class files
did not have references for the R$drawable.foo identifiers.
This results in a crash.

So to fix the problem we need to start looking at the
`@(LibraryResourceDirectories)` to see if we have any
AndroidManifest.xml files. We should then process these
directories to ensure that the R.java files are created.
The finally compiled into the final applications.