[Xamarin.Android.Build.Tasks] SlidingMenuExample project now fails to build against 15.4 with dozens of javac errors 'package foo does not exist'
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=58727

In commit 0667a2b6 we neglected to take into account that any
embedded libraries or environment files would be deleted when
the zip was extracted. Commit ea6b9b45 fixed the issue for the
Environment files. This commit will fix the issue for the .jar
files.

We also added a new test so this doesn't happen in future.