[xa-prep-tasks] HashFileContents should exclude line endings (#747)
Fix `$(_CrossOutputDirTop)`, which is missing a closing parenthesis,
which angers MSBuild on Windows. (Why it isn't a problem on macOS...)

Update the `<HashFileContents>` task to exclude line endings.

`.gitattributes` changes the line endings for MSBuild project files,
which causes a different hash to be generated on Windows than macOS.
This means that Windows can't use a prebuilt bundle, as the generated
bundle name depends on the content hash of files which differ on
Windows; see commit 97f08f78.

Avoid this problem by reading the files into a `MemoryStream`,
ignoring `\r` and `\n` characters, then hash the `MemoryStream`
contents. This allows both macOS and Windows to produce the same
content hash values, regardless of line endings, which in turn
will allow a future Windows build to use the macOS-built bundle.