[build] Update `$(AndroidApiLevel)`=26
Commit 759fd472 [broke the **macOS+msbuild** build][0]:

[0]: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android-msbuild/387/

	error XA0001: Unsupported or invalid $(TargetFrameworkVersion) value of 'v7.1'. Please update your Project Options.

The cause of the problem is the `<ResolveSdks/>` task: if both
`$(AndroidApiLevel)` and `$(TargetFrameworkVersion)` are set, then
`$(AndroidApiLevel)` takes precedence, and we set
`$(TargetFrameworkVersion)` to the value that corresponds to
`$(AndroidApiLevel)`.

Commit 759fd472 updated `$(AndroidFrameworkVersion)` to v8.0, so
"normal" builds (`make all`) will only build `Mono.Android.dll` for
v8.0. However, 759fd472 forgot to update `$(AndroidApiLevel)` to 26 --
the API level corresponding to v8.0 -- which resulted in a mismatch
when building e.g. `Xamarin.Android.NUnitLite.csproj`:
`$(AndroidApiLevel)` was 25, which implies/requires
`$(TargetFrameworkVersion)`=v7.1, but we never built v7.1, which
resulted in the above XA0001 error.

Fix the problem by updating `$(AndroidApiLevel)` to default to
`$(AndroidLatestApiLevel)` -- which is 26 as of 759fd472 -- to help
avoid this mismatch again.